chore(parser): move more client interfaces to openapi
