+++ src/openApi/v2/parser/getOperationResults.ts
@@ -1,4 +1,4 @@
-import type { Model, OperationResponse } from '../../../types/client';
+import type { Model, OperationResponse } from '../../common/interfaces/client';
 
 const areEqual = (a: Model, b: Model): boolean => {
     const equal = a.type === b.type && a.base === b.base && a.template === b.template;
