+++ src/openApi/v2/parser/getServices.ts
@@ -1,6 +1,6 @@
-import type { Service } from '../../../types/client';
 import type { Config } from '../../../types/config';
 import { unique } from '../../../utils/unique';
+import type { Service } from '../../common/interfaces/client';
 import type { OpenApi } from '../interfaces/OpenApi';
 import { getOperation } from './getOperation';
 import { getOperationParameters } from './getOperationParameters';
