+++ src/utils/handlebars.ts
@@ -2,6 +2,7 @@ import camelCase from 'camelcase';
 import Handlebars from 'handlebars/runtime';
 import { EOL } from 'os';
 
+import type { Model, OperationParameter, Service } from '../openApi';
 import templateClient from '../templates/client.hbs';
 import angularGetHeaders from '../templates/core/angular/getHeaders.hbs';
 import angularGetRequestBody from '../templates/core/angular/getRequestBody.hbs';
@@ -86,7 +87,7 @@ import partialTypeInterface from '../templates/partials/typeInterface.hbs';
 import partialTypeIntersection from '../templates/partials/typeIntersection.hbs';
 import partialTypeReference from '../templates/partials/typeReference.hbs';
 import partialTypeUnion from '../templates/partials/typeUnion.hbs';
-import type { Client, Model, OperationParameter, Service } from '../types/client';
+import type { Client } from '../types/client';
 import type { Config } from '../types/config';
 import { enumKey, enumName, enumUnionType, enumValue } from './enum';
 import { escapeName } from './escapeName';
